requirements.txt
# Environment or requirements
fastapi
uvicorn
pillow
numpy
torch
transformers
pydantic
tensorflow